Refactoring of model to entity
