+++ src/main/java/cz/inspire/common/entity/NastaveniJsonEntity.java
@@ -1,6 +1,10 @@
-package cz.inspire.common.model;
+package cz.inspire.common.entity;
 
-import jakarta.persistence.*;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,20 +18,12 @@ import java.util.Objects;
 @Getter
 @NoArgsConstructor
 @Entity
-@Table(name="mena")
-public class MenaEntity {
+@Table(name="nastaveni_json")
+public class NastaveniJsonEntity {
     @Id
-    private String id;
+    private String key;
     @Column
-    private String kod;
-    @Column
-    private String vycetka;
-    @Column
-    private int kodnum;
-    @Column
-    private int zaokrouhleniHotovost;
-    @Column
-    private int zaokrouhleniKarta;
+    private String value;
 
     @Override
     public final boolean equals(Object o) {
@@ -36,8 +32,8 @@ public class MenaEntity {
         Class<?> oEffectiveClass = o instanceof HibernateProxy ? ((HibernateProxy) o).getHibernateLazyInitializer().getPersistentClass() : o.getClass();
         Class<?> thisEffectiveClass = this instanceof HibernateProxy ? ((HibernateProxy) this).getHibernateLazyInitializer().getPersistentClass() : this.getClass();
         if (thisEffectiveClass != oEffectiveClass) return false;
-        MenaEntity obj = (MenaEntity) o;
-        return getId() != null && Objects.equals(getId(), obj.getId());
+        NastaveniJsonEntity obj = (NastaveniJsonEntity) o;
+        return getKey() != null && Objects.equals(getKey(), obj.getKey());
     }
 
     @Override
