+++ src/main/java/cz/inspire/common/mapper/HeaderMapper.java
@@ -1,7 +1,7 @@
 package cz.inspire.common.mapper;
 
 import cz.inspire.common.dto.HeaderDto;
-import cz.inspire.common.model.HeaderEntity;
+import cz.inspire.common.entity.HeaderEntity;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "cdi")
