+++ src/main/java/cz/inspire/common/mapper/NastaveniMapper.java
@@ -1,6 +1,6 @@
 package cz.inspire.common.mapper;
 
-import cz.inspire.common.model.NastaveniEntity;
+import cz.inspire.common.entity.NastaveniEntity;
 import cz.inspire.common.dto.NastaveniDto;
 import org.mapstruct.Mapper;
