+++ src/main/java/cz/inspire/common/repository/MenaRepository.java
@@ -1,6 +1,6 @@
 package cz.inspire.common.repository;
 
-import cz.inspire.common.model.MenaEntity;
+import cz.inspire.common.entity.MenaEntity;
 import org.apache.deltaspike.data.api.EntityRepository;
 import org.apache.deltaspike.data.api.Query;
 import org.apache.deltaspike.data.api.Repository;
