+++ src/main/java/cz/inspire/common/repository/NastaveniRepository.java
@@ -1,6 +1,6 @@
 package cz.inspire.common.repository;
 
-import cz.inspire.common.model.NastaveniEntity;
+import cz.inspire.common.entity.NastaveniEntity;
 import org.apache.deltaspike.data.api.EntityRepository;
 import org.apache.deltaspike.data.api.Repository;
