+++ src/main/java/cz/inspire/common/service/NastaveniJsonService.java
@@ -1,6 +1,6 @@
 package cz.inspire.common.service;
 
-import cz.inspire.common.model.NastaveniJsonEntity;
+import cz.inspire.common.entity.NastaveniJsonEntity;
 import cz.inspire.common.repository.NastaveniJsonRepository;
 import cz.inspire.exception.SystemException;
 import jakarta.ejb.CreateException;
