+++ src/test/java/RepositoryTests/CommonServer/HeaderRepositoryTest.java
@@ -1,7 +1,7 @@
 package RepositoryTests.CommonServer;
 
 import cz.inspire.thesis.data.EntityManagerProducer;
-import cz.inspire.common.model.HeaderEntity;
+import cz.inspire.common.entity.HeaderEntity;
 import cz.inspire.common.repository.HeaderRepository;
 import jakarta.persistence.*;
 import org.apache.deltaspike.cdise.api.CdiContainer;
